move delegate_asset macro to shared macros module
Relocate `delegate_asset!` from b20_asset/logic/mod.rs to the crate's
macros module (its common home alongside decode_precompile_call), so
future versions reference it from one place. It stays Asset-trait-
specific; recursion uses the $crate::macros path so the use-imported
macro resolves regardless of call site.
